[IMP] stock_planning, stock_planning_mrp, stock_planning_procurement_generated_by_plan

+++ stock_planning/wizard/wiz_stock_planning.py
@@ -2,7 +2,7 @@
 ##############################################################################
 # For copyright and license notices, see __openerp__.py file in root directory
 ##############################################################################
-from openerp import models, fields, api, _
+from openerp import models, fields, api, exceptions, _
 from dateutil.relativedelta import relativedelta
 
 
@@ -22,7 +22,7 @@ class WizStockPlanning(models.TransientModel):
         default=lambda self: fields.Date.context_today(self),
         help='Date from which the interval starts counting days')
     days = fields.Integer(
-        'Days interval', required=True,
+        'Days interval', required=True, default=1,
         help='Increase number of days starting from the date from')
     to_date = fields.Date('To date', required=True,
                           help='Deadline for calculating periods')
@@ -35,6 +35,10 @@ class WizStockPlanning(models.TransientModel):
     product = fields.Many2one(
         'product.product', 'Product',
         help='Enter this field if you want to filter by product')
+    locations = fields.Many2many(
+        comodel_name='stock.location', relation='rel_stock_planning_location',
+        column1='wiz_stock_planning_id', column2='locations_id',
+        string='Locations')
 
     @api.multi
     def calculate_stock_planning(self):
@@ -42,7 +46,14 @@ class WizStockPlanning(models.TransientModel):
         planning_obj = self.env['stock.planning']
         move_obj = self.env['stock.move']
         proc_obj = self.env['procurement.order']
+        if self.days < 1:
+            raise exceptions.Warning(
+                _('Error!: Increase number of days must be greater than zero'))
         cond = [('company', '=', self.company.id)]
+        if self.locations:
+            cond.append(('location', 'in', self.locations.ids))
+        if self.product:
+            cond.append(('product', '=', self.product.id))
         planning = planning_obj.search(cond)
         planning.unlink()
         fdate = self.from_date
@@ -56,20 +67,22 @@ class WizStockPlanning(models.TransientModel):
             self.company, fdate, category=self.category,
                 template=self.template, product=self.product):
             if move.location_id.usage == 'internal':
-                product_datas = self._find_product_in_table(
-                    product_datas, move.product_id, move.location_id,
-                    move.warehouse_id)
+                if (not self.locations or
+                    (self.locations and move.location_id.id in
+                     self.locations.ids)):
+                    product_datas = self._find_product_in_table(
+                        product_datas, move.product_id, move.location_id)
             if move.location_dest_id.usage == 'internal':
-                product_datas = self._find_product_in_table(
-                    product_datas, move.product_id,
-                    move.location_dest_id, move.warehouse_id)
-        states = ('confirmed', 'exception')
+                if (not self.locations or
+                    (self.locations and move.location_dest_id.id in
+                     self.locations.ids)):
+                    product_datas = self._find_product_in_table(
+                        product_datas, move.product_id, move.location_dest_id)
         for procurement in proc_obj._find_procurements_from_stock_planning(
-            self.company, fdate, states, category=self.category,
-                template=self.template, product=self.product):
+            self.company, fdate, category=self.category,
+                template=self.template, product=self.product, periods=True):
             product_datas = self._find_product_in_table(
-                product_datas, procurement.product_id,
-                procurement.location_id, procurement.warehouse_id)
+                product_datas, procurement.product_id, procurement.location_id)
         self._generate_stock_planning(product_datas)
         return {'name': _('Stock Planning'),
                 'type': 'ir.actions.act_window',
@@ -78,25 +91,19 @@ class WizStockPlanning(models.TransientModel):
                 'res_model': 'stock.planning',
                 }
 
-    def _find_product_in_table(self, product_datas, product, location,
-                               warehouse):
+    def _find_product_in_table(self, product_datas, product, location):
         found = False
         for data in product_datas:
             datos_array = product_datas[data]
             dproduct = datos_array['product']
             dlocation = datos_array['location']
-            dwarehouse = datos_array['warehouse']
             if dproduct.id == product.id and dlocation.id == location.id:
                 found = True
-                if not dwarehouse and warehouse:
-                    product_datas[data].update({'warehouse': warehouse})
-                break
         if not found:
             my_vals = {'product': product,
                        'location': location,
-                       'warehouse': warehouse,
                        }
-            ind = product.id + location.id + (warehouse.id or 0)
+            ind = product.id + location.id
             product_datas[(ind)] = my_vals
         return product_datas
 
@@ -113,14 +120,6 @@ class WizStockPlanning(models.TransientModel):
                         'product': datos_array['product'].id}
                 if from_date:
                     vals['from_date'] = from_date
-                if datos_array['warehouse']:
-                    vals['warehouse'] = datos_array['warehouse'].id
-                else:
-                    cond = [('company_id', '=', self.company.id),
-                            ('lot_stock_id', '=', datos_array['location'].id)]
-                    warehouses = self.env['stock.warehouse'].search(cond)
-                    if warehouses:
-                        vals['warehouse'] = warehouses[0].id
                 planning_obj.create(vals)
                 from_date = fields.Date.to_string(
                     fields.Date.from_string(fdate) + relativedelta(days=1))
